Working to update the TorSpider script.
I'm working to update the TorSpider script, preparing for multi-node
support and meanwhile fixing database communication so that the spiders
can work ceaselessly without having to wait for updates to the database.

This is the first commit, and it's not near complete.
